Move image sorting function back to pipeline subroutines collection.
